Fix 3 test failures, replace shortcut with global EveryScene KSPAddon
Test fixes:
- MainCameraExists: check PlanetariumCamera in tracking station
  (Camera.main is null there)
- CommittedRecordingsHaveValidData: skip 0-point tree root recordings
- EveryRecordingHasResolvablePart: try raw name before underscore
  split, demote to warning (showcase parts legitimately unresolvable)

Shortcut rework:
- New TestRunnerShortcut KSPAddon(EveryScene) with Input.GetKeyDown
  in Update() — reliable across all scenes (IMGUI Event.current was
  swallowed by KSP in Flight/KSC)
- Removed per-scene shortcut handlers from ParsekUI, ParsekFlight,
  ParsekKSC, and ParsekTrackingStation
- Removed duplicate test runner UI from ParsekTrackingStation (global
  addon covers it)
