+++ Source/Parsek/ParsekFlight.cs
@@ -643,7 +643,6 @@ namespace Parsek
                 ui.DrawActionsWindowIfOpen(windowRect);
                 ui.DrawSettingsWindowIfOpen(windowRect);
                 ui.DrawSpawnControlWindowIfOpen(windowRect);
-                ui.HandleTestRunnerShortcut();
                 ui.DrawTestRunnerWindowIfOpen(windowRect, this);
             }
         }
