+++ Source/Parsek/ParsekKSC.cs
@@ -111,7 +111,6 @@ namespace Parsek
             ui.DrawRecordingsWindowIfOpen(windowRect);
             ui.DrawActionsWindowIfOpen(windowRect);
             ui.DrawSettingsWindowIfOpen(windowRect);
-            ui.HandleTestRunnerShortcut();
             ui.DrawTestRunnerWindowIfOpen(windowRect, this);
         }
